+++ src/base/fldini.f90
@@ -97,7 +97,6 @@ integer          f_dim, kiflux, kbflux
 
 character(len=80) :: name, f_name
 
-type(gas_mix_species_prop) sasp
 type(var_cal_opt) :: vcopt
 
 !===============================================================================
@@ -680,112 +679,6 @@ do f_id = 0, nfld - 1
     call field_set_key_struct_var_cal_opt(f_id, vcopt)
   endif
 enddo
-
-! Copy field physical properties of species into the field structure
-
-if (ippmod(igmix).ge.0) then
-
-  call cs_parameters_define_field_key_gas_mix
-
-  do f_id = 0, nfld - 1
-
-    call field_get_name (f_id, name)
-
-    if (name.eq.'y_o2') then
-
-      sasp%mol_mas= 0.032d0
-      sasp%cp = 930.d0
-      sasp%vol_dif = 19.70d0
-      sasp%mu_a = 5.086522d-8
-      sasp%mu_b = 5.512391d-6
-      sasp%lambda_a = 6.2d-5
-      sasp%lambda_b = 8.1d-3
-      sasp%muref = 1.919d-5
-      sasp%lamref = 0.0244d0
-      sasp%trefmu = 273.d0
-      sasp%treflam = 273.d0
-      sasp%smu = 139.d0
-      sasp%slam = 240.d0
-
-      call field_set_key_struct_gas_mix_species_prop(f_id, sasp)
-
-    else if (name.eq.'y_n2') then
-
-      sasp%mol_mas = 0.028d0
-      sasp%cp = 1042.d0
-      sasp%vol_dif = 19.70d0
-      sasp%mu_a = 4.210130d-8
-      sasp%mu_b = 5.494348d-6
-      sasp%lambda_a = 6.784141d-5
-      sasp%lambda_b = 5.564317d-3
-      sasp%muref = 1.663d-5
-      sasp%lamref = 0.0242d0
-      sasp%trefmu = 273.d0
-      sasp%treflam = 273.d0
-      sasp%smu = 107.d0
-      sasp%slam = 150.d0
-
-      call field_set_key_struct_gas_mix_species_prop(f_id, sasp)
-
-    else if (name.eq.'y_he') then
-
-      sasp%mol_mas = 0.004d0
-      sasp%cp = 5194.d0
-      sasp%vol_dif = 2.67d0
-      sasp%mu_a = 18.5752d-6
-      sasp%mu_b = 0.0d0
-      sasp%lambda_a = 0.144d0
-      sasp%lambda_b = 0.0d0
-      sasp%muref = 1.874d-5
-      sasp%lamref = 0.647d0
-      sasp%trefmu = 273.d0
-      sasp%treflam = 273.d0
-      sasp%smu = 78.d0
-      sasp%slam = 78.d0
-
-      call field_set_key_struct_gas_mix_species_prop(f_id, sasp)
-
-    else if (name.eq.'y_h2') then
-
-      sasp%mol_mas = 0.002d0
-      sasp%cp = 14560.d0
-      sasp%vol_dif = 6.12d0
-      sasp%mu_a = 1.93d-9
-      sasp%mu_b = 8.40d-6
-      sasp%lambda_a = 4.431d-4
-      sasp%lambda_b = 5.334d-2
-      sasp%muref = 8.411d-6
-      sasp%lamref = 0.0168d0
-      sasp%trefmu = 273.d0
-      sasp%treflam = 273.d0
-      sasp%smu = 97.d0
-      sasp%slam = 120.d0
-
-      call field_set_key_struct_gas_mix_species_prop(f_id, sasp)
-
-    else if (name.eq.'y_h2o_g') then
-
-      sasp%mol_mas = 0.018d0
-      sasp%cp = 2060.d0
-      sasp%vol_dif = 13.10d0
-      sasp%mu_a = 3.8496d-8
-      sasp%mu_b = 8.2997d-6
-      sasp%lambda_a = 7.6209d-5
-      sasp%lambda_b = 0.016949d0
-      sasp%muref = 1.12d-5
-      sasp%lamref = 0.0181d0
-      sasp%trefmu = 350.d0
-      sasp%treflam = 300.d0
-      sasp%smu = 1064.d0
-      sasp%slam = 2200.d0
-
-      call field_set_key_struct_gas_mix_species_prop(f_id, sasp)
-
-    endif
-
-  enddo
-
-endif
 
 ! Check if scalars are buoyant and set n_buoyant_scal accordingly.
 ! It is then used in tridim to update buoyant scalars and density in U-P loop